+++ src/preload/index.ts
@@ -682,7 +682,7 @@ const api: ApplicationAPI = {
 
   writeToClipboard: (text: string) => ipcRenderer.invoke('clipboard-write-text', text),
   openPath: (path: string) => ipcRenderer.invoke('open-path', path),
-  openUrlInWindow: (url: string) => ipcRenderer.invoke('open-url-in-window', url),
+  openUrlInWindow: (url: string, title?: string) => ipcRenderer.invoke('open-url-in-window', url, title),
 
   // System logs
   getSystemLogs: (fromId?: number, limit?: number, levels?: SystemLogLevel[]) => ipcRenderer.invoke('get-system-logs', fromId, limit, levels),